Changes made after review of PR #20.
* Updated access modifier on SaveSCMPluginConfigRequest constructor
* Move SCMIntegrationPluginConfiguration from an inner class to a new dto's package
